backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pymongo==4.6.1
pymysql==1.1.0
cryptography==41.0.7
sqlalchemy==2.0.25
websockets==12.0
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
google-generativeai==0.3.2
groq==1.0.0

# FIXED: supported by Python 3.12
numpy==1.26.4

# Graph Analysis (Optional - for advanced clustering)
networkx==3.2.1
python-louvain==0.16
